backend/requirements.txt
flask==3.1.1
flask-restful==0.3.10
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==23.0.0
pytest==8.3.5